test: createGameSession 이관으로 인한 단위테스트케이스 수정
